+++ amazon/core/src/main/kotlin/org/http4k/connect/amazon/core/xml.kt
@@ -1,6 +1,6 @@
 package org.http4k.connect.amazon.core
 
-import org.http4k.core.Request
+import org.http4k.core.Body
 import org.http4k.core.Response
 import org.w3c.dom.Document
 import org.w3c.dom.Node
@@ -9,8 +9,8 @@ import org.xml.sax.InputSource
 import java.io.StringReader
 import javax.xml.parsers.DocumentBuilderFactory
 
-fun Response.xmlDoc(): Document = documentBuilderFactory().parse(body.stream)
-fun Request.xmlDoc(): Document = documentBuilderFactory().parse(body.stream)
+fun Response.xmlDoc() = body.xmlDoc()
+fun Body.xmlDoc(): Document = documentBuilderFactory().parse(stream)
 
 private fun documentBuilderFactory() =
     DocumentBuilderFactory.newInstance()