+++ amazon/s3/fake/src/main/kotlin/org/http4k/connect/amazon/s3/FakeS3.kt
@@ -50,9 +50,9 @@ class FakeS3(
         isS3 bind routes(
             pathBasedCopyKey(buckets, bucketContent, clock),
             pathBasedBucketGetKey(buckets, bucketContent),
-            pathBasedBucketPostKey(buckets, bucketContent),
+            pathBasedBucketPostKey(buckets, bucketContent, clock),
             pathBasedBucketPutKey(buckets, bucketContent, clock),
-            pathBasedBucketDeleteKey(buckets, bucketContent),
+            pathBasedBucketDeleteKey(buckets, bucketContent, clock),
             pathBasedBucketHeadKey(buckets, bucketContent),
             globalListObjectsV2(buckets, bucketContent),
             globalHeadBucket(buckets),
@@ -64,9 +64,9 @@ class FakeS3(
             bucketHeadBucket(buckets),
             bucketHeadKey(buckets, bucketContent),
             bucketGetKey(buckets, bucketContent),
-            bucketPostKey(buckets, bucketContent),
+            bucketPostKey(buckets, bucketContent, clock),
             bucketPutKey(buckets, bucketContent, clock),
-            bucketDeleteKey(buckets, bucketContent),
+            bucketDeleteKey(buckets, bucketContent, clock),
             bucketPutBucket(buckets),
             bucketDeleteBucket(buckets),
             bucketListObjectsV2(buckets, bucketContent)