+++ amazon/s3/fake/src/main/kotlin/org/http4k/connect/amazon/s3/endpoints/common.kt
@@ -2,17 +2,22 @@ package org.http4k.connect.amazon.s3.endpoints
 
 import org.http4k.connect.amazon.s3.BucketKeyContent
 import org.http4k.connect.amazon.s3.S3Error
+import org.http4k.connect.amazon.s3.TestingHeaders.X_HTTP4K_LAST_MODIFIED
 import org.http4k.connect.storage.Storage
 import org.http4k.core.Body
 import org.http4k.core.ContentType.Companion.APPLICATION_XML
+import org.http4k.core.Headers
 import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status.Companion.FORBIDDEN
 import org.http4k.core.Status.Companion.NOT_FOUND
 import org.http4k.core.with
+import org.http4k.lens.LastModified
 import org.http4k.routing.asRouter
 import org.http4k.template.HandlebarsTemplates
 import org.http4k.template.viewModel
+import java.time.Clock
+import java.time.Instant
 
 
 fun Request.subdomain(buckets: Storage<Unit>): String =
@@ -30,20 +35,21 @@ val s3ErrorLens by lazy {
 
 const val GLOBAL_BUCKET = "unknown"
 
-fun invalidBucketNameResponse() = Response(NOT_FOUND)
+internal fun invalidBucketNameResponse() = Response(NOT_FOUND)
     .with(s3ErrorLens of S3Error("NoSuchBucket", message = "The resource you requested does not exist"))
 
-fun invalidBucketKeyResponse() = Response(NOT_FOUND)
+internal fun invalidBucketKeyResponse() = Response(NOT_FOUND)
     .with(s3ErrorLens of S3Error("NoSuchKey", message = "The resource you requested does not exist"))
 
-fun invalidObjectStateResponse() = Response(FORBIDDEN)
+internal fun invalidObjectStateResponse() = Response(FORBIDDEN)
     .with(s3ErrorLens of S3Error("InvalidObjectState", message = "Object is in an invalid state"))
 
 internal val excludedObjectHeaders = setOf(
     "authorization",
     "x-forwarded-host",
     "x-amz-content-sha256",
-    "x-amz-date"
+    "x-amz-date",
+    "x-amz-tagging",
 )
 
 internal fun getHeadersWithoutXHttp4kPrefix(it: BucketKeyContent) =
@@ -51,3 +57,10 @@ internal fun getHeadersWithoutXHttp4kPrefix(it: BucketKeyContent) =
 
 // TODO may be overlooked that `queries` router only passes if the query has a value
 fun queryPresent(name: String) = { req: Request -> req.queries(name).isNotEmpty() }.asRouter("Query present: $name")
+
+// TODO may want to consider adding to http4k-core routing.kt
+val otherwise = { _: Request -> true }.asRouter("Catch-all")
+
+internal fun lastModified(headers: Headers, clock: Clock) = headers
+    .firstOrNull { it.first == X_HTTP4K_LAST_MODIFIED }?.second?.let { LastModified.parse(it) }?.value?.toInstant()
+    ?: Instant.now(clock)