+++ amazon/s3/fake/src/main/kotlin/org/http4k/connect/amazon/s3/endpoints/postKey.kt
@@ -1,14 +1,13 @@
 package org.http4k.connect.amazon.s3.endpoints
 
 import org.http4k.connect.amazon.core.firstChild
+import org.http4k.connect.amazon.core.model.RfcTimestamp
 import org.http4k.connect.amazon.core.sequenceOfNodes
 import org.http4k.connect.amazon.core.text
 import org.http4k.connect.amazon.core.xmlDoc
 import org.http4k.connect.amazon.s3.BucketKeyContent
-import org.http4k.connect.amazon.s3.action.RestoreObject
 import org.http4k.connect.amazon.s3.model.BucketKey
 import org.http4k.connect.amazon.s3.model.BucketName
-import org.http4k.connect.amazon.s3.model.RestoreTier
 import org.http4k.connect.amazon.s3.replaceHeader
 import org.http4k.connect.amazon.s3.requiresRestore
 import org.http4k.connect.amazon.s3.storageClass
@@ -20,8 +19,10 @@ import org.http4k.core.Status.Companion.OK
 import org.http4k.routing.bind
 import org.http4k.routing.path
 import org.http4k.routing.routes
+import java.time.Clock
+import java.time.Duration
 
-fun bucketPostKey(buckets: Storage<Unit>,  bucketContent: Storage<BucketKeyContent>) =
+fun bucketPostKey(buckets: Storage<Unit>,  bucketContent: Storage<BucketKeyContent>, clock: Clock) =
     "/{bucketKey:.+}" bind POST to routes(
         queryPresent("restore") bind { request ->
             restoreObject(
@@ -29,12 +30,13 @@ fun bucketPostKey(buckets: Storage<Unit>,  bucketContent: Storage<BucketKeyConte
                 bucketContent,
                 BucketName.of(request.subdomain(buckets)),
                 BucketKey.of(request.path("bucketKey")!!),
-                request
+                request,
+                clock
             )
         }
     )
 
-fun pathBasedBucketPostKey(buckets: Storage<Unit>,  bucketContent: Storage<BucketKeyContent>) =
+fun pathBasedBucketPostKey(buckets: Storage<Unit>,  bucketContent: Storage<BucketKeyContent>, clock: Clock) =
     "/{bucketName}/{bucketKey:.+}" bind POST to routes(
         queryPresent("restore") bind { request ->
             restoreObject(
@@ -42,7 +44,8 @@ fun pathBasedBucketPostKey(buckets: Storage<Unit>,  bucketContent: Storage<Bucke
                 bucketContent,
                 BucketName.of(request.path("bucketName")!!),
                 BucketKey.of(request.path("bucketKey")!!),
-                request
+                request,
+                clock
             )
         }
     )
@@ -53,26 +56,23 @@ private fun restoreObject(
     bucketContent: Storage<BucketKeyContent>,
     bucket: BucketName,
     bucketKey: BucketKey,
-    request: Request
+    request: Request,
+    clock: Clock
 ): Response {
-    val doc = request.xmlDoc().getElementsByTagName("RestoreRequest").sequenceOfNodes().first()
+    val doc = request.body.xmlDoc().getElementsByTagName("RestoreRequest").sequenceOfNodes().first()
 
-    val data = RestoreObject(
-        key = bucketKey,
-        days = doc.firstChild("Days")!!.text().toInt(),
-        tier = doc.firstChild("GlacierJobParameters")
-            ?.firstChild("Tier")
-            ?.text()
-            ?.let { RestoreTier.valueOf(it) }
-    )
+    val days = doc.firstChild("Days")!!.text().toLong()
+    val expires = clock.instant() + Duration.ofDays(days)
 
     if (buckets[bucket.value] == null) return invalidBucketNameResponse()
-    val obj = bucketContent["$bucket-${data.key}"] ?: return invalidBucketKeyResponse()
+    val obj = bucketContent["$bucket-$bucketKey"] ?: return invalidBucketKeyResponse()
     if (!obj.storageClass().requiresRestore()) return invalidObjectStateResponse()
 
     // instantly restore the object, regardless of restore tier
-    // TODO could use clock to approximate restore time
-    bucketContent["$bucket-${data.key}"] = obj.replaceHeader("x-amz-restore", "ongoing-request=\"false\"")
+    bucketContent["$bucket-$bucketKey"] = obj.replaceHeader(
+        "x-amz-restore",
+        "ongoing-request=\"false\", expiry-date=\"${RfcTimestamp.of(expires)}\""
+    )
 
     return Response(OK)
 }